Reconcile the changes for CSV input with the type cleanup changes

+++ geocsv.cpp
@@ -9,7 +9,7 @@
 #include "csv.hpp"
 #include "milo/dtoa_milo.h"
 
-void parse_geocsv(std::vector<struct serialization_state> &sst, std::string fname, int layer, std::string layername) {
+void parse_geocsv(std::vector<struct serialization_state> &sst, std::string fname, size_t layer, std::string layername) {
 	FILE *f = fopen(fname.c_str(), "r");
 	if (f == NULL) {
 		perror(fname.c_str());
@@ -68,7 +68,7 @@ void parse_geocsv(std::vector<struct serialization_state> &sst, std::string fnam
 		double lon = atof(line[loncol].c_str());
 		double lat = atof(line[latcol].c_str());
 
-		long long x, y;
+		long x, y;
 		projection->project(lon, lat, 32, &x, &y);
 		drawvec dv;
 		dv.push_back(draw(VT_MOVETO, x, y));
